requirements.txt
django
django-heroku
gunicorn
django-markdownx
dj-database-url
